backend/requirements.txt
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
dnspython==2.5.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pymongo==4.6.1
python-dotenv==1.0.1
Werkzeug==3.0.1
